[2/3] monkey-patch PetConsts correctly
`LuaScripts/main.lua` is failing to run because the fixes for the missing `PetConsts` values doesn't work with the current set of downloaded files.

This diff fixes the monkey-patching by:
- Running it during the invocation of `Loader:main()` instead of during the loading of `Loader.lua`. This reduces the risk of the fixes being overwritten by something else.
- Making it add missing values per-key instead of doing all-or-nothing on `PetConsts`. This was a problem because the most recent version of `LuaEnum.lua` includes `PetConsts`, but is missing some keys that are referenced in the docs.

After this fix, I was able to successfully run `LuaScripts/main.lua` and the results of that run are the next diff in this stack.